Attempt at modularizing Xenomorph.
Segregated the dependencies into their own modules, and moved code
around to make it compile. The number of imports required, as a result,
to make the specs compile is not very satisfactory, but I did not want
to overload the commit.
