--- src/test/scala/xenomorph/json/JTypeSpec.scala
@@ -1,51 +1,0 @@
-/*
- * Copyright (C) 2017 Kris Nuttycombe
- * All rights reserved.
- *
- * This file is part of the Scala Xenomorph library.
- *
- * GNU Lesser General Public License Usage
- * This file may be used under the terms of the GNU Lesser
- * General Public License version 3.0 as published by the Free Software
- * Foundation and appearing in the file LICENSE included in the
- * packaging of this file.  Please review the following information to
- * ensure the GNU Lesser General Public License version 3.0 requirements
- * will be met: https://www.gnu.org/licenses/lgpl-3.0.txt
- */
-package xenomorph
-
-import org.specs2._
-import org.scalacheck._
-
-import xenomorph.json.JType._
-import xenomorph.json.ToJson._
-import xenomorph.json.FromJson._
-import xenomorph.scalacheck.ToGen._
-
-import xenomorph.samples._
-
-class JTypeSpec extends Specification with org.specs2.ScalaCheck {
-  def is = s2"""
-  Serialization of values to JSON should
-    serialize a value to JSON $toJson
-    read a value from JSON $fromJson
-    round-trip values produced by a generator $gen
-  """
-
-  def toJson = {
-    val result = Person.schema.toJson(person) 
-    result.toString must_== """{"roles":[{"administrator":{"subordinateCount":0,"department":"windmill-tilting"}}],"birthDate":20147028000,"name":"Kris Nuttycombe"}"""
-  }
-
-  def fromJson = {
-    val result = Person.schema.toJson(person) 
-    Person.schema.fromJson(result).toOption must_== Some(person)
-  }
-
-  def gen = {
-    implicit val arbPerson = Arbitrary(Person.schema.toGen)
-    prop(
-      (p: Person) => Person.schema.fromJson(Person.schema.toJson(p)).toOption must_== Some(p)
-    )
-  }
-}
